refactor: change ProtocolController to new standard
